instrument: refactor, format, and organize code by integration
