+++ internal/instrument/instrument.go
@@ -302,10 +302,10 @@ func addInFunctionCode(list []dst.Stmt, tc *typechecker.TypeChecker, conf config
 				appendStmt = false
 				out = append(out, ginMiddleware(stmt))
 			}
-			if isEcho(stmt) {
+			if isEchoV4(stmt) {
 				out = append(out, stmt)
 				appendStmt = false
-				out = append(out, echoMiddleware(stmt))
+				out = append(out, echoV4Middleware(stmt))
 			}
 
 			// Recurse when there is a function literal on the RHS of the assignment.
@@ -388,32 +388,6 @@ func instrumentReturn(stmt *dst.ReturnStmt) *dst.ReturnStmt {
 	return wrapSqlReturnCall(stmt)
 }
 
-func buildFunctionLiteralHandlerCode(name dst.Expr, funLit *dst.FuncLit) []dst.Stmt {
-	//check if magic comment is attached to first line
-	if len(funLit.Body.List) > 0 {
-		decs := funLit.Body.List[0].Decorations().Start
-		for _, v := range decs.All() {
-			if strings.HasPrefix(v, dd_startinstrument) {
-				log.Println("already instrumented")
-				return funLit.Body.List
-			}
-		}
-	}
-	// get name of request var
-	requestName := funLit.Type.Params.List[1].Names[0].Name
-	newLines := buildFunctionInstrumentation(name, requestName)
-	funLit.Body.List = append(newLines, funLit.Body.List...)
-	return funLit.Body.List
-}
-
-func analyzeExpressionForHandlerLiteral(funLit *dst.FuncLit, tc *typechecker.TypeChecker) bool {
-	// check the parameters, see if they match
-	inputParams := funLit.Type.Params.List
-	return len(inputParams) == 2 &&
-		tc.OfType(inputParams[0].Type, "net/http.ResponseWriter") &&
-		tc.OfType(inputParams[1].Type, "*net/http.Request")
-}
-
 const (
 	dd_startinstrument = "//dd:startinstrument"
 	dd_endinstrument   = "//dd:endinstrument"
@@ -431,334 +405,6 @@ func hasLabel(label string, decs []string) bool {
 		}
 	}
 	return false
-}
-
-func analyzeStmtForRequestClient(stmt *dst.AssignStmt) (string, bool) {
-	// looking for
-	// 	req, err := http.NewRequestWithContext(context.Background(), http.MethodPost, "localhost:8080", strings.NewReader(os.Args[1]))
-	// has 2 return values (*http.Request and error)
-	// function named NewRequestWithContext
-	if len(stmt.Lhs) == 2 &&
-		len(stmt.Rhs) == 1 {
-		if fun, ok := stmt.Rhs[0].(*dst.CallExpr); ok {
-			if iden, ok := fun.Fun.(*dst.Ident); ok {
-				if iden.Name == "NewRequestWithContext" && iden.Path == "net/http" {
-					if iden, ok := stmt.Lhs[0].(*dst.Ident); ok {
-						return iden.Name, true
-					}
-				}
-			}
-		}
-	}
-	return "", false
-}
-
-func wrapFromAssign(stmt *dst.AssignStmt, tc *typechecker.TypeChecker) bool {
-	return wrapHandlerFromAssign(stmt, tc) || wrapClientFromAssign(stmt, tc)
-
-}
-
-func wrapGRPC(stmt *dst.AssignStmt) {
-	/*
-		//dd:startwrap
-		s := grpc.NewServer(opt1, opt2, orchestrion.GRPCStreamServerInterceptor(), orchestrion.GRPCUnaryServerInterceptor())
-		//dd:endwrap
-	*/
-
-	/*
-		//dd:startwrap
-		c, err := grpc.Dial(target, opt1, orchestrion.GRPCStreamClientInterceptor(), orchestrion.GRPCUnaryClientInterceptor())
-		//dd:endwrap
-	*/
-	if !(len(stmt.Lhs) <= 2 && len(stmt.Rhs) == 1) {
-		return
-	}
-	wrap := func(targetName string, opts ...string) {
-		if fun, ok := stmt.Rhs[0].(*dst.CallExpr); ok {
-			if iden, ok := fun.Fun.(*dst.Ident); ok {
-				if !(iden.Name == targetName && iden.Path == "google.golang.org/grpc") {
-					return
-				}
-				stmt.Decorations().Start.Append(dd_startwrap)
-				stmt.Decorations().End.Append("\n", dd_endwrap)
-				for _, opt := range opts {
-					fun.Args = append(fun.Args,
-						&dst.CallExpr{Fun: &dst.Ident{Name: opt, Path: "github.com/datadog/orchestrion/instrument"}},
-					)
-				}
-			}
-		}
-	}
-	wrap("NewServer", "GRPCStreamServerInterceptor", "GRPCUnaryServerInterceptor")
-	wrap("Dial", "GRPCStreamClientInterceptor", "GRPCUnaryClientInterceptor")
-}
-
-func wrapHandlerFromAssign(stmt *dst.AssignStmt, tc *typechecker.TypeChecker) bool {
-	/*
-		s = &http.Server{
-			//dd:startwrap
-			Handler: orchestrion.WrapHandler(handler),
-			//dd:endwrap
-		}
-	*/
-	if !(len(stmt.Lhs) == 1 && len(stmt.Rhs) == 1) {
-		return false
-	}
-	if uexpr, ok := stmt.Rhs[0].(*dst.UnaryExpr); ok {
-		if x, ok := uexpr.X.(*dst.CompositeLit); ok {
-			t, ok := x.Type.(*dst.Ident)
-			if !(ok && t.Path == "net/http" && t.Name == "Server") {
-				return false
-			}
-			for _, e := range x.Elts {
-				if hasLabel(dd_startwrap, e.Decorations().Start.All()) {
-					return false
-				}
-				if kve, ok := e.(*dst.KeyValueExpr); ok {
-					k, ok := kve.Key.(*dst.Ident)
-					if !(ok && k.Name == "Handler" && tc.OfType(k, "net/http.Handler")) {
-						continue
-					}
-					kve.Decorations().Start.Append(dd_startwrap)
-					kve.Decorations().End.Append("\n", dd_endwrap)
-					kve.Value = &dst.CallExpr{
-						Fun:  &dst.Ident{Name: "WrapHandler", Path: "github.com/datadog/orchestrion/instrument"},
-						Args: []dst.Expr{kve.Value},
-					}
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
-func wrapClientFromAssign(stmt *dst.AssignStmt, tc *typechecker.TypeChecker) bool {
-	/*
-		//dd:startwrap
-		c = orchestrion.WrapHTTPClient(client)
-		//dd:endwrap
-	*/
-	if !(len(stmt.Lhs) == 1 && len(stmt.Rhs) == 1) {
-		return false
-	}
-	iden, ok := stmt.Lhs[0].(*dst.Ident)
-	if !(ok && tc.OfType(iden, "*net/http.Client")) {
-		return false
-	}
-	stmt.Decorations().Start.Append(dd_startwrap)
-	stmt.Decorations().End.Append("\n", dd_endwrap)
-	stmt.Rhs[0] = &dst.CallExpr{
-		Fun:  &dst.Ident{Name: "WrapHTTPClient", Path: "github.com/datadog/orchestrion/instrument"},
-		Args: []dst.Expr{stmt.Rhs[0]},
-	}
-	return true
-}
-
-func wrapSqlReturnCall(stmt *dst.ReturnStmt) *dst.ReturnStmt {
-	/*
-		//dd:startwrap
-		return sql.Open("postgres", "somepath")
-		//dd:endwrap
-
-		//dd:startwrap
-		return sql.OpenDB(connector)
-		//dd:endwrap
-	*/
-	for _, expr := range stmt.Results {
-		fun, ok := expr.(*dst.CallExpr)
-		if !ok {
-			continue
-		}
-		if wrapSqlCall(fun) {
-			stmt.Decorations().Start.Append(dd_startwrap)
-			stmt.Decorations().Before = dst.NewLine
-			stmt.Decorations().End.Append("\n", dd_endwrap)
-		}
-
-	}
-	return stmt
-}
-
-func wrapSqlOpenFromAssign(stmt *dst.AssignStmt) bool {
-	/*
-		//dd:startwrap
-		db, err = sql.Open("postgres", "somepath")
-		//dd:endwrap
-
-		//dd:startwrap
-		db = sql.OpenDB(connector)
-		//dd:endwrap
-	*/
-
-	rhs := stmt.Rhs[0]
-	fun, ok := rhs.(*dst.CallExpr)
-	if !ok {
-		return false
-	}
-	if wrapSqlCall(fun) {
-		stmt.Decorations().Start.Append(dd_startwrap)
-		stmt.Decorations().End.Append("\n", dd_endwrap)
-		return true
-	}
-	return false
-}
-
-func wrapSqlCall(call *dst.CallExpr) bool {
-	f, ok := call.Fun.(*dst.Ident)
-	if !(ok && f.Path == "database/sql" && (f.Name == "Open" || f.Name == "OpenDB")) {
-		return false
-	}
-	f.Path = "github.com/datadog/orchestrion/instrument"
-	return true
-}
-
-func wrapHandlerFromExpr(stmt *dst.ExprStmt, tc *typechecker.TypeChecker) bool {
-	/*
-		//dd:startwrap
-		http.Handle("/handle", orchestrion.WrapHandler(handler))
-		//dd:endwrap
-
-		//dd:startwrap
-		http.HandleFunc("/handle", orchestrion.WrapHandlerFunc(handler))
-		//dd:endwrap
-
-		//dd:startwrap
-		s.Handle("/handle", orchestrion.WrapHandler(handler))
-		//dd:endwrap
-
-		//dd:startwrap
-		s.HandleFunc("/handle", orchestrion.WrapHandlerFunc(handler))
-		//dd:endwrap
-	*/
-	wrap := func(fun *dst.CallExpr, name string) bool {
-		var wrapper string
-		switch name {
-		case "Handle":
-			wrapper = "WrapHandler"
-		case "HandleFunc":
-			wrapper = "WrapHandlerFunc"
-		default:
-			return false
-		}
-		fun.Decorations().Start.Append(dd_startwrap)
-		fun.Decorations().End.Append("\n", dd_endwrap)
-		fun.Args[1] = &dst.CallExpr{
-			Fun:  &dst.Ident{Name: wrapper, Path: "github.com/datadog/orchestrion/instrument"},
-			Args: []dst.Expr{fun.Args[1]},
-		}
-		return true
-	}
-	if fun, ok := stmt.X.(*dst.CallExpr); ok && len(fun.Args) == 2 {
-		switch f := fun.Fun.(type) {
-		case *dst.SelectorExpr:
-			if tc.OfType(f.X, "*net/http.ServeMux") || tc.OfType(f.X, "net/http.ServeMux") {
-				return wrap(fun, f.Sel.Name)
-			}
-		case *dst.Ident:
-			if f.Path == "net/http" {
-				return wrap(fun, f.Name)
-			}
-		}
-	}
-	return false
-}
-
-func buildRequestClientCode(requestName string) dst.Stmt {
-	/*
-		//dd:startinstrument
-		if req != nil {
-			req = req.WithContext(Report(req.Context(), EventCall, "url", req.URL, "method", req.Method))
-			req = InsertHeader(req)
-			defer Report(req.Context(), EventReturn, "url", req.URL, "method", req.Method)
-		}
-		//dd:endinstrument
-
-	*/
-	return &dst.IfStmt{
-		Cond: &dst.BinaryExpr{
-			X:  &dst.Ident{Name: requestName},
-			Op: token.NEQ,
-			Y:  &dst.Ident{Name: "nil"},
-		},
-		Body: &dst.BlockStmt{
-			List: []dst.Stmt{
-				&dst.AssignStmt{
-					Lhs: []dst.Expr{&dst.Ident{Name: requestName}},
-					Tok: token.ASSIGN,
-					Rhs: []dst.Expr{
-						&dst.CallExpr{
-							Fun: &dst.SelectorExpr{
-								X:   &dst.Ident{Name: requestName},
-								Sel: &dst.Ident{Name: "WithContext"},
-							},
-							Args: []dst.Expr{
-								&dst.CallExpr{
-									Fun: &dst.Ident{Name: "Report", Path: "github.com/datadog/orchestrion/instrument"},
-									Args: []dst.Expr{
-										&dst.CallExpr{Fun: &dst.SelectorExpr{
-											X:   &dst.Ident{Name: requestName},
-											Sel: &dst.Ident{Name: "Context"},
-										}},
-										&dst.Ident{Name: "EventCall", Path: "github.com/datadog/orchestrion/instrument/event"},
-										&dst.BasicLit{Kind: token.STRING, Value: `"name"`},
-										&dst.SelectorExpr{
-											X:   &dst.Ident{Name: requestName},
-											Sel: &dst.Ident{Name: "URL"},
-										},
-										&dst.BasicLit{Kind: token.STRING, Value: `"verb"`},
-										&dst.SelectorExpr{
-											X:   &dst.Ident{Name: requestName},
-											Sel: &dst.Ident{Name: "Method"},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-				&dst.AssignStmt{
-					Lhs: []dst.Expr{&dst.Ident{Name: requestName}},
-					Tok: token.ASSIGN,
-					Rhs: []dst.Expr{
-						&dst.CallExpr{
-							Fun:  &dst.Ident{Name: "InsertHeader", Path: "github.com/datadog/orchestrion/instrument"},
-							Args: []dst.Expr{&dst.Ident{Name: requestName}},
-						},
-					},
-				},
-				&dst.DeferStmt{Call: &dst.CallExpr{
-					Fun: &dst.Ident{Name: "Report", Path: "github.com/datadog/orchestrion/instrument"},
-					Args: []dst.Expr{
-						&dst.CallExpr{Fun: &dst.SelectorExpr{
-							X:   &dst.Ident{Name: requestName},
-							Sel: &dst.Ident{Name: "Context"},
-						}},
-						&dst.Ident{Name: "EventReturn", Path: "github.com/datadog/orchestrion/instrument/event"},
-						&dst.BasicLit{Kind: token.STRING, Value: `"name"`},
-						&dst.SelectorExpr{
-							X:   &dst.Ident{Name: requestName},
-							Sel: &dst.Ident{Name: "URL"},
-						},
-						&dst.BasicLit{Kind: token.STRING, Value: `"verb"`},
-						&dst.SelectorExpr{
-							X:   &dst.Ident{Name: requestName},
-							Sel: &dst.Ident{Name: "Method"},
-						},
-					},
-				},
-				},
-			},
-		},
-		Decs: dst.IfStmtDecorations{
-			NodeDecs: dst.NodeDecs{
-				Before: dst.NewLine,
-				Start:  dst.Decorations{dd_startinstrument},
-				After:  dst.NewLine,
-				End:    dst.Decorations{"\n", dd_endinstrument},
-			},
-		},
-	}
 }
 
 func addInit(decl *dst.FuncDecl) *dst.FuncDecl {
@@ -787,26 +433,6 @@ func addInit(decl *dst.FuncDecl) *dst.FuncDecl {
 		},
 	}
 
-	decl.Body.List = append(newLines, decl.Body.List...)
-	return decl
-}
-
-func addCodeToHandler(decl *dst.FuncDecl) *dst.FuncDecl {
-	//check if magic comment is attached to first line
-	if len(decl.Body.List) > 0 {
-		decs := decl.Body.List[0].Decorations().Start
-		for _, v := range decs.All() {
-			if strings.HasPrefix(v, dd_startinstrument) {
-				log.Println("already instrumented")
-				return decl
-			}
-		}
-	}
-	// get name of request var
-	requestName := decl.Type.Params.List[1].Names[0].Name
-	newLines := buildFunctionInstrumentation(
-		&dst.BasicLit{Kind: token.STRING, Value: `"` + decl.Name.Name + `"`},
-		requestName)
 	decl.Body.List = append(newLines, decl.Body.List...)
 	return decl
 }
@@ -907,103 +533,6 @@ func dup(in dst.Expr) dst.Expr {
 	}
 }
 
-func reportHandlerFromAssign(stmt *dst.AssignStmt, tc *typechecker.TypeChecker, conf config.Config) {
-	// check for function literal that is a handler
-	for pos, expr := range stmt.Rhs {
-		if compLit, ok := expr.(*dst.CompositeLit); ok {
-			for _, v := range compLit.Elts {
-				if kv, ok := v.(*dst.KeyValueExpr); ok {
-					if funLit, ok := kv.Value.(*dst.FuncLit); ok {
-						if analyzeExpressionForHandlerLiteral(funLit, tc) {
-							// get the name from the field
-							funLit.Body.List = buildFunctionLiteralHandlerCode(kv.Key, funLit)
-						}
-					}
-				}
-			}
-		}
-		if funLit, ok := expr.(*dst.FuncLit); ok {
-			if analyzeExpressionForHandlerLiteral(funLit, tc) {
-				// get the name from the lhs in the same position -- if it's not there, exit, code isn't going to compile
-				if len(stmt.Lhs) <= pos {
-					break
-				}
-				funLit.Body.List = buildFunctionLiteralHandlerCode(stmt.Lhs[pos], funLit)
-			}
-		}
-	}
-}
-
-func reportHandlerFromExpr(stmt *dst.ExprStmt, tc *typechecker.TypeChecker, conf config.Config) {
-	// might be something we have to recurse on if it's a closure?
-	if call, ok := stmt.X.(*dst.CallExpr); ok {
-		// check if this is a handler func
-		switch funLit := call.Fun.(type) {
-		case *dst.FuncLit:
-			if analyzeExpressionForHandlerLiteral(funLit, tc) {
-				funLit.Body.List = buildFunctionLiteralHandlerCode(nil, funLit)
-			}
-		}
-		// check if any of the parameters is a function literal
-		var prevExpr dst.Expr
-		for _, v := range call.Args {
-			if funLit, ok := v.(*dst.FuncLit); ok {
-				// check for function literal that is a handler
-				if analyzeExpressionForHandlerLiteral(funLit, tc) {
-					funLit.Body.List = buildFunctionLiteralHandlerCode(prevExpr, funLit)
-				}
-			}
-			prevExpr = v
-		}
-	}
-}
-
-func reportHandlerFromDecl(decl *dst.FuncDecl, tc *typechecker.TypeChecker, conf config.Config) *dst.FuncDecl {
-	if conf.HTTPMode != "report" {
-		return decl
-	}
-	// check the parameters, see if they match
-	inputParams := decl.Type.Params.List
-	if len(inputParams) == 2 &&
-		tc.OfType(inputParams[0].Type, "net/http.ResponseWriter") &&
-		tc.OfType(inputParams[1].Type, "*net/http.Request") {
-		decl = addCodeToHandler(decl)
-	}
-	return decl
-}
-
-func isGin(stmt *dst.AssignStmt) bool {
-	rhs := stmt.Rhs[0]
-	f, ok := funcIdent(rhs)
-	return ok && f.Path == "github.com/gin-gonic/gin" && (f.Name == "New" || f.Name == "Default")
-}
-
-func ginMiddleware(got *dst.AssignStmt) dst.Stmt {
-	iden, ok := got.Lhs[0].(*dst.Ident)
-	if !ok {
-		return nil
-	}
-	stmt := useMiddleware(iden.Name, "GinMiddleware")
-	wrap(stmt)
-	return stmt
-}
-
-func isEcho(stmt *dst.AssignStmt) bool {
-	rhs := stmt.Rhs[0]
-	f, ok := funcIdent(rhs)
-	return ok && f.Path == "github.com/labstack/echo/v4" && f.Name == "New"
-}
-
-func echoMiddleware(got *dst.AssignStmt) dst.Stmt {
-	iden, ok := got.Lhs[0].(*dst.Ident)
-	if !ok {
-		return nil
-	}
-	stmt := useMiddleware(iden.Name, "EchoV4Middleware")
-	wrap(stmt)
-	return stmt
-}
-
 func funcIdent(e dst.Expr) (*dst.Ident, bool) {
 	call, ok := e.(*dst.CallExpr)
 	if !ok {
@@ -1036,7 +565,7 @@ func useMiddleware(pkg, middleware string) *dst.ExprStmt {
 	return stmt
 }
 
-func wrap(stmt dst.Stmt) {
+func wrap(stmt dst.Node) {
 	stmt.Decorations().Start.Append(dd_startwrap)
 	stmt.Decorations().End.Append("\n", dd_endwrap)
 }
