+++ internal/instrument/uninstrument.go
@@ -35,7 +35,7 @@ var unwrappers = []func(n dst.Node) bool{
 // removing instrumentation that adds code.
 var removers = []func(stmt dst.Stmt) bool{
 	removeGin,
-	removeEcho,
+	removeEchoV4,
 }
 
 func UninstrumentFile(name string, r io.Reader, conf config.Config) (io.Reader, error) {
@@ -180,181 +180,6 @@ func removeStartEndInstrument(list []dst.Stmt) []dst.Stmt {
 		}
 	}
 	return list
-}
-
-// unwrapClient unwraps client, to be used in dst.Inspect.
-// Returns true to continue the traversal, false to stop.
-func unwrapClient(n dst.Node) bool {
-	s, ok := n.(*dst.AssignStmt)
-	if !ok {
-		return true
-	}
-	ce, ok := s.Rhs[0].(*dst.CallExpr)
-	if !ok {
-		return true
-	}
-	cei, ok := ce.Fun.(*dst.Ident)
-	if !ok {
-		return true
-	}
-	if cei.Path == "github.com/datadog/orchestrion/instrument" && strings.HasPrefix(cei.Name, "WrapHTTPClient") {
-		s.Rhs[0] = ce.Args[0]
-		return false
-	}
-	return true
-}
-
-// unwrapHandlerExpr unwraps handler expressions, to be used in dst.Inspect.
-// Returns true to continue the traversal, false to stop.
-func unwrapHandlerExpr(n dst.Node) bool {
-	es, ok := n.(*dst.ExprStmt)
-	if !ok {
-		return true
-	}
-	f, ok := es.X.(*dst.CallExpr)
-	if !ok {
-		return true
-	}
-	if len(f.Args) > 1 {
-		if ce, ok := f.Args[1].(*dst.CallExpr); ok {
-			if cei, ok := ce.Fun.(*dst.Ident); ok {
-				if cei.Path == "github.com/datadog/orchestrion/instrument" &&
-					strings.HasPrefix(cei.Name, "WrapHandler") {
-					// This catches both WrapHandler *and* WrapHandlerFunc
-					f.Args[1] = ce.Args[0]
-					return false
-				}
-			}
-		}
-	}
-	return true
-}
-
-// unwrapHandlerAssign unwraps handler assignements, to be used in dst.Inspect.
-// Returns true to continue the traversal, false to stop.
-func unwrapHandlerAssign(n dst.Node) bool {
-	// TODO: Implement me
-	return false
-}
-
-func unwrapSqlExpr(n dst.Node) bool {
-	es, ok := n.(*dst.ExprStmt)
-	if !ok {
-		return true
-	}
-	f, ok := es.X.(*dst.CallExpr)
-	if !ok {
-		return true
-	}
-	id, ok := f.Fun.(*dst.Ident)
-	if !ok {
-		return true
-	}
-	if id.Path == "github.com/datadog/orchestrion/instrument" &&
-		(id.Name == "Open" || id.Name == "OpenDB") {
-		id.Path = "database/sql"
-		return true
-	}
-	return true
-}
-
-func unwrapSqlAssign(n dst.Node) bool {
-	as, ok := n.(*dst.AssignStmt)
-	if !ok {
-		return true
-	}
-	f, ok := as.Rhs[0].(*dst.CallExpr)
-	if !ok {
-		return true
-	}
-	id, ok := f.Fun.(*dst.Ident)
-	if !ok {
-		return true
-	}
-	if id.Path == "github.com/datadog/orchestrion/instrument" &&
-		(id.Name == "Open" || id.Name == "OpenDB") {
-		id.Path = "database/sql"
-		return true
-	}
-	return true
-}
-
-func unwrapSqlReturn(n dst.Node) bool {
-	rs, ok := n.(*dst.ReturnStmt)
-	if !ok {
-		return true
-	}
-	for _, expr := range rs.Results {
-		fun, ok := expr.(*dst.CallExpr)
-		if !ok {
-			continue
-		}
-		f, ok := fun.Fun.(*dst.Ident)
-		if !(ok && f.Path == "github.com/datadog/orchestrion/instrument" &&
-			(f.Name == "Open" || f.Name == "OpenDB")) {
-			continue
-		}
-		f.Path = "database/sql"
-	}
-	return true
-}
-
-// unwrapGRPC unwraps grpc server and client, to be used in dst.Inspect.
-// Returns true to continue the traversal, false to stop.
-func unwrapGRPC(n dst.Node) bool {
-	s, ok := n.(*dst.AssignStmt)
-	if !ok {
-		return true
-	}
-	ce, ok := s.Rhs[0].(*dst.CallExpr)
-	if !ok {
-		return true
-	}
-	cei, ok := ce.Fun.(*dst.Ident)
-	if !ok {
-		return true
-	}
-	if cei.Path != "google.golang.org/grpc" || !(cei.Name == "Dial" || cei.Name == "NewServer") || len(ce.Args) == 0 {
-		return true
-	}
-	removeLast := func(args []dst.Expr, targetFunc string) []dst.Expr {
-		if len(args) == 0 {
-			return args
-		}
-		lastArg := args[len(args)-1]
-		lastArgExp, ok := lastArg.(*dst.CallExpr)
-		if !ok {
-			return args
-		}
-		fun, ok := lastArgExp.Fun.(*dst.Ident)
-		if !ok {
-			return args
-		}
-		if !(fun.Path == "github.com/datadog/orchestrion/instrument" && fun.Name == targetFunc) {
-			return args
-		}
-		return args[:len(args)-1]
-	}
-	removable := []string{
-		"GRPCUnaryServerInterceptor",
-		"GRPCStreamServerInterceptor",
-		"GRPCUnaryClientInterceptor",
-		"GRPCStreamClientInterceptor",
-	}
-	for _, opt := range removable {
-		ce.Args = removeLast(ce.Args, opt)
-	}
-	return true
-}
-
-// removeGin returns whether a statement corresponds to orchestrion's Gin-middleware registration
-func removeGin(stmt dst.Stmt) bool {
-	return removeUseMiddleware(stmt, "GinMiddleware")
-}
-
-// removeEcho returns whether a statement corresponds to orchestrion's Echo-middleware registration
-func removeEcho(stmt dst.Stmt) bool {
-	return removeUseMiddleware(stmt, "EchoV4Middleware")
 }
 
 func removeUseMiddleware(stmt dst.Stmt, name string) bool {
