+++ samples/server/echov4.go
@@ -11,7 +11,7 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func echoServer() {
+func echoV4Server() {
 	r := echo.New()
 	r.GET("/ping", func(c echo.Context) error {
 		return c.JSON(http.StatusOK, map[string]any{
